[BP] MB-62596: Update getNumItemsForCursor to be constant time
This is commit squashes the multiple development commits and critical
fixes which were later found. There is no need to try and backport
many patches when logically there is only one commit for the fix.

4 commits make up the change

https://review.couchbase.org/c/kv_engine/+/212289
https://review.couchbase.org/c/kv_engine/+/214270

With the following two fixes found from system testing (MB-62961)

https://review.couchbase.org/c/kv_engine/+/214599
https://review.couchbase.org/c/kv_engine/+/214270

There are additional changes made.

1) As the code counts checkpoint_start the backported code failed
the MB_58342 unit test. This was missed in the orginal code because
the unit test had been removed. Both this backport and master
versions are fixed by a "fudge" of the items_remaining when the
cursor is on empty with only the checkpoint_start ahead.

See CheckpointCursor::getNumItems for the code change and this 
change itself is tracked by MB-65280

2) To mitigate against there being an issue in the accounting the
subtraction which generates the items_remaining value is guarded
against underflow, the guard is not to throw/crash the system but
to force a return value of 0.

See CheckpointCursor::getNumItems for the code change.

Now follows the original commit message describing the fix:

Change the CheckpointManager/Cursor code so they work together
to compute items available.

CheckpointManager tracks "totalItems" which is a monotonically
increasing counter of items put into Checkpoints, e.g. every
mutation,  checkpoint_start/end and queueVBState, i.e. anything
which should be accounted in getNumItemsForCursor (so not the
queue_op::empty item).

This can be thought of as a number-line where totalItems is the
end of the line.

Cursor now tracks it's position on this number-line and we can
obtain items remaining by subtracting the position from the
totalItems.

Change-Id: I174a14cfc1451308931126c0c2aea3dd3bd897fd
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/223288
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Jim Walker <jim@couchbase.com>