[Cleanup] Remove dynamic checks for valgrind in unit tests
Remove the dynamic code path in order to check if the test program
is run under valgrind. The motivation for adding these checks was
to speed up the CV runs when we used valgrind as part of that.
These days we've replaced the CV valgrind runs with TSan, ASan
and UBSan.

One may still run the binary under valgrind if one ever wanted
to do that, but the iteration numbers would be higher and deeper
testing would happen.

Change-Id: Ia5ac17f536f9679f652b4c173f17f12893ea9ebe
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/223351
Reviewed-by: Mohammad Zaeem <mohammad.zaeem@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@couchbase.com>