MB-64937: Fix secondaryWarmup atomic initialisation
The method getSecondaryWarmup() reads the atomic with relaxed memory
ordering. This is wrong - we need to use acquire, to be able to see the
Warmup object in an initialised state.

We've seen this trip TSan with a message indicating that the Warmup
object is still being initialised when the pointer is seen from another
thread.

When the Warmup object is created, we update the atomic pointer with
release semantics. The standard says:

    All writes in the current thread are visible in other threads that
    acquire the same atomic variable.

The key is acquire - to see the writes in the initialisation thread, we
need to use acquire semantics when reading the pointer.

I've moved the Synchronized<unique_ptr<T>> + atomic<T*> into a new
struct called SynchronizedInitOncePtr, which correctly does rel+acq.

Change-Id: Icef141bef8fd5ad2442aae80d9d5e9ba13c7f342
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/223390
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Reviewed-by: Mohammad Zaeem <mohammad.zaeem@couchbase.com>