MB-65320: Add --sort to mcstat
Allow for specifying --sort on the command line to let mcstat
spool the entire output in memory and sort it before dumping
it to standard output. Note you won't get any output if there
is an error on the server causing the connection to be dropped
or another error occurs.

Change-Id: I0d5adb3da21fb7f8c69bb22c4b4ebb5a4cf91077
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/223473
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>