MB-65403: Disconnect DCP producer when "stuck"
Add code to DCP producer to allow the enforcement of a new rule.
The rule is that when a producer is paused due to flow
control, if it makes no progress (no ack is received) for
a configurable amount of time (default 720s) the producer is
disconnected.

The rule is optionally enforced by the use of a regular expression
that is applied against a DCP producer's name. By default the
regex is an empty string, and this means no match is even attempted,
and the feature is disabled.

The configuration is passed using memcached.json to simplify the
management of the feature in mixed-mode clusters.

To ease the passing of the regex inside the JSON (where we would
need to take care to escape certain patterns) the value is expected
to be base64 encoded, removing the need to escape certain character
sequences.

Change-Id: I9ce48b05b87d0784b1e456a6ca75a3b59107a21c
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/224051
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Jim Walker <jim@couchbase.com>