MB-65468: Release pagerSemaphore when filter is empty
ItemPager does not release tokens it acquired from the pagerSemaphore if
the filter is empty. No PagingVisitors are scheduled in such a case
either to release the tokens. This puts the ItemPager in a
non-functional state where future runs will not be able to acquire the
tokens required to schedule PagingVisitor/s.

The filter can be empty if there are no vbuckets in a state we can
evict from. This can happen when a bucket is flushed where all vbuckets
are put into the dead state. Since this is not a state we evict from,
the filter will return empty and so no PagingVisitors will be scheduled
as expected. However, we do not release the tokens in such a case and so
all future executions will not be able to acquire from the semaphore.

Change-Id: I96b2f2bdcc324761b1b40c8e9334b3e01af621fa
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/223981
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Well-Formed: Restriction Checker