MB-65450: reset weak_ptr on the correct memory arena
The weak_ptr which is owned by the ConnectionIface is destructed
by the core. Even though the weak_ptr doesn't own the ConnHandler
it can still be responsible for freeing the memory that was
allocated to the orginal shared_ptr.

Even though we think that the shared_ptr destruction should be
what frees memory, it turns out that any weak_ptrs can still
keep the memory referenced and be responsbile for deallocation.

The following blog offers a much more detailed explanation:

https://www.cppstories.com/2017/12/weakptr-memory/

To fix, ConnectionIface must switch to the allocators arena and
reset.

This fix has been verfied using an enhanced version of
CB_ARENA_MALLOC_VERIFY_DEALLOC_CLIENT which validates that thing freed
in the global domain were allocated by the global domain (not in this
case).

Change-Id: I19153d947f18be97b5d6d5ae03b77b7c6f073eed
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/224127
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>