MB-65450: Fix bucket_logger use of allocation guards
logWithContext needs to move the alloc guard to the outer scope so
that the catch block if entered will deallocate memory correctly
(this fix was already applied to logInner).

It was then noted that capture the current engine must also happen
before the alloc guard, else it will always be null.

Change-Id: I8c19fde398324ca9608ace8b390adf4e9c3bc712
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/224129
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>