MB-60861: Always call logWithContext from BucketLogger
Changed the implementation of the logging routine used by the
EP_LOG_* macros to call logWithContext with the bucket and conn_id
passed as a JSON object.

This effectively converts all of the old-style logging:

  <TS> (Bucket) Message
  <TS> 123: (Bucket) Message

To the new JSON logging:

  <TS> Message {"bucket":"Bucket"}
  <TS> Message {"conn_id":123,"bucket":"Bucket"}

Without having to change the macro that is being used.

Change-Id: I535b473d5ad0b0b8e0dcf6f06f5d760e70fd449a
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/221688
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>