MB-65742/MB-51689: Make highPreparedSeqno non-monotonic
When sending a disk snapshot we set the HPS to that of the active. We
then hit a monotonic violation when processing prepares as it will have
a HPS <= to the one initially set.

This is a temporary fix that should be reverted when a fix for MB-51689
is merged.

Change-Id: If74abf54fdf5282e5603fc28dfe380dbc87566c0
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/225069
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>