MB-65742/MB-51689: Fix monotonic violation in HPS
Prepares on disk are de-duped when a subsequent commit is received for
an item. Therefore when we backfill from disk, the replica doesn't
receive the Prepares corresponding to Durable Writes.

We do trasmit the HPS for the Disk Snapshot in the SnapshotMarker. The
Disk Snapshot could have a mix of committed items & prepared items and
therefore can't be set directly on the checkpoint when the
snapshotMarker is received. We should update the HPS on the checkpoint
when the full snapshot is processed (on processing the mutation with seqno
SnapEnd).

There is an new field introduced in Checkpoint - snapHighPrepareSeqno,
which signifies the high prepare seqno of all the mutations in a given
snapshot. It's set when a snapshotMarker V2.2 is received.

https://review.couchbase.org/c/kv_engine/+/225069 is reverted too as a
part of this patch where we remove the monotonic constraint on the hps.

The test introduced as a part of
https://review.couchbase.org/c/kv_engine/+/225099, which repro-ed the
original issue seen in the MB-51689 has been updated too to make the
correct assertions.

Change-Id: I65b77963f834540989326cacebfc38cfb91cd13b
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/225175
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>