MB-64001: Expand MutationResponse ahead of adding CacheTransferStream
CacheTransferStream will reuse MutationResponse for queuing "mutations"
for transfer.

Later when it comes to the DcpProducer sending the correct message for
an item from cache, we will not be sending a DcpMutation. Instead a
different opcode will be used (so we can strongly differentiate the
items from cache vs items from traditional DCP).

Here the change updates MutationResponse so that the Event which is
currently derived from the Item can be overriden. This will allow
CacheTransferStream to create and queue MutationResponse objest and
ensure that the Event type is never Mutation, Deletion, Expiration,
Prepare or Abort.

Change-Id: Ib68dea48205da8c57a5d61d6ff470c1987a3c7a7
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/229888
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>