MB-57405: Update MagmaDbStats atomically at compaction
Previously, we used to update the stats when the first SST was compacted
(Magma persisted the update immediately). Instead, we should update the
stats when the compaction is complete, in the same transaction as
clearing the DroppedCollections.

We retry the compaction when the droppedCollections has the record of
the collection and that retains the collection count, so clearing both
the droppedCollection and decrementing the stats should be valid.

Updated the ErrorInjection tests with tests which fail halfway in the
compaction.

Change-Id: I70ed5938e946d3c8f276079daab14b8c5b0fb8e5
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/225261
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>