MB-64993: Add support for unknown config parameters
On 7.2, we crash when attempting to create a bucket with an unknown
config key. This behaviour was changed in 7.6 to allow bucket creation
with unknown configs.

This patch will ignore any unknown config values and only set valid
configuration values. This allows us to set config values before
carrying out an upgrade which supports that config.

Change-Id: I9513961e54af3e712c7a4c7078e5b55f96119b29
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/225428
Well-Formed: Restriction Checker
Tested-by: Faizan Alam <faizan.alam@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>