MB-66011: Don't reset the privilege context to "no bucket"...
... as part of authentication, but leave it in the bucket currently
selected if the client currently connected still have access to
the bucket.

This isn't really visible for "normal" authentication as the client
typically perform a single authentication cycle and at that time the
connected buket is no bucket. The problem is visible when one use JWT
and OAUTHBEARER and the token expire. At this time the server will
return "AuthStale" which would cause a client to send a new token. The
client shouldn't have to send a new "select bucket" command as it is
already connected to the bucket and the privilege context should
reflect that.

Change-Id: Idf3804c18e1eeaea1d88261c5424f200ce9a15e7
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/225519
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>