MB-65737: Fail flush if a document has seqno less or equal to persisted
When flushing documents to disk, we must never proceed if any document
has a seqno number smaller than or equal to the current persisted seqno
on disk. Doing so would lead to overwriting data on disk. To prevent
data loss, this patch fails any flush that contains a document with a
seqno number less than or equal to the current persisted seqno on disk.

Change-Id: I4c6a5cde02ce1b1076958d9aa1d70baf8098dba2
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/225114
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>