MB-65719: Enforce that MagmaKVStoreConfig::store is always valid
The member is a pointer to the owning MagmaKVStore instance, thus by
logic always not-null.

A T& would be the proper type for encoding that semantic. Problem is,
the code that constructs the various KVStore/KVStoreConfig
implementations makes the conversion to T& non-trivial, so I'm avoiding
that refactor at this stage.

Change-Id: Id58e2b3e3f0a7db4e7098b15cbe840e3c325a7f6
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/226065
Reviewed-by: Mohammad Zaeem <mohammad.zaeem@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>