MB-65849: Fusion stats fail gracefully
When gathering fusion stats, any error from magma triggers an exception
in EPE. The result is that the connection is disconnected.
Thus, first reason for the change is that in general we prefer to avoid
disconnections where possible.

Second reason for the change is related to the specific case of
KVStoreInvalid under Fusion.

In its procedures ns_server initializes fusion KVStore(s) (by
SetVBucketState(vbid)) and then it polls fusion stats for the vbucket.

Given that KVStore creation at SetVBucketState is asynchronous, then
ns_server might end up polling stats for some vbucket before the
related KVStore is created on disk.

In that case magma returns KVStoreInvalid to the caller. KV on its side
throws and closes the connection.

Given that in that case KVStoreInvalid is a benign/transient failure,
then by this patch KV avoids disconnections and just returns
NotMyVbucket (MCBP) to the caller.

Any unexpected error other than KVStoreInvalid translates to Einternal
(MCBP).

Change-Id: Ic84d08fd3130dca2466024eb8398a9cd5c84a286
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/226299
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>