MB-63679: Expose aggregated 'fusion active_guest_volumes' stats
KV exposed the per-vbucket variant already:
  'fusion active_guest_volumes <vbid>'

This patch allows to omit the <vbid> arg to retrieve the aggregated
list of all volumes for all vbuckets in a bucket.

The format of the response is unchanged, volumes are always returned in
a json array. The aggregate call just merges all vbuckets' list into an
array of unique elements.

Change-Id: I870b4aa486d598565354f11539510f63bf2830b1
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/226462
Reviewed-by: Mohammad Zaeem <mohammad.zaeem@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>