MB-65770: Fix intermittent failure of ItemPagerWithAllDeadVBuckets
We currently fill memory with relatively large items which causes
flakiness due to a small 2.5MiB quota.

As a fix, we add smaller items to reduce the chances of individual
items having a large affect on memUsage. Added this functionality to
populateVbsUntil() as other tests have similar behaviour such as the one
in https://review.couchbase.org/c/kv_engine/+/221757 which can utilise
this functionality.

Change-Id: I597ba6ae1a8d2c907fb299d0fe41fa420a1150d8
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/226217
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>