MB-66262: Refactor: (client) getMeta should throw...
...an exception if the command fails like the other methods
in MemcachedConnection. If one wants to test "error scenarios"
and other return values one should use the execute() method
and look at the returned object. The helper methods in
MemcachedConnection object is intended to simplify the
boilerplate code to write unit tests (as the error checks
would be hidden inside the method for the common case)

Change-Id: If841494217c4b96e9ffa823dff6cb5e324a79ef6
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/226272
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>