MB-64001: Switch to ActiveStream after CacheTransfer
If the stream-request end-seqno is above the start-seqno, then a
CacheTransfer should "swap" over to ActiveStream and continue producing
DCP messages.

This is achieved by the CacheTransferStream looking at the
stream-request and queueing a "meta" message into the readyQueue.

The DcpProducer will perform a "swap" of the stream when it processes
the CacheTransferToActiveStream message. Note that two reasons for
using a meta-message.

1) Lock-inversion. Trying to get the CacheTransferStream to swap itself
is difficult, primarily we cannot access the DcpProducer streams
container as it is usually locked when we're executing a
CacheTransferStream function

2). Ordering. The CacheTransferToActiveStream will be ordered after all
CacheTransfer messages, and as such the switch to ActiveStream cannot
jump ahead of data that must be processed before it.

Change-Id: I71d525bd199922d64b6cdc1cf92c05343cda2505
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/232769
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>