Create a spdlog wrapper in the core
Refactors a lot of the BucketLogger implementation into
cb::logger::Logger. The new Logger class sinks messages into a base
splogger in the same way BucketLogger used to and retains the same
properties (shadowed log() to control formatting allocation) and so on.

It also adds the withContext methods.

We now create a Logger instance which wraps the "base" global spdlogger.

This will allow us to build a generic prefixed logger like BucketLogger
but in the core.

Change-Id: I90afd169f8c497b30aa9f090bbb47dbde7800d51
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/226361
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>