Add PrefixLogger which appends some predefined context to logs
Refactor more of BucketLogger into the core. PrefixLogger instances can
be used to create a logger scoped to a component, like we already do
with EP-Engine and DCP ConnHandler logging.

Note PrefixLogger, like other loggers, needs to be registered if the
logger needs to support dynamic change of verbosity. In the future, we
could have a (one) PrefixLogger instance in the core which reads conn_id
and other thread_local state and automatically prefixes logs with that.

Now BucketLogger is left only with the functionality of prepending the
engine and conn_id.

Change-Id: I4531991e047b41bfda54fcaec2e432fb900f1613
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/226362
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>