MB-66186: Avoid race when the chk remover signals before atomic store
The nature of the race is that the woken chk remover can complete and
call signal() before the removersToComplete.store(#chk-removers)
happens, in which case the signal() method fail with assertion that the
removersToComplete is 0.

We need to update removersToComplete before we wake up the tasks. I've
changed the interface of wakeUpChkRemoversAndGetNotified to take the
expected number, as this will allow the chkRemovers to be resized in the
future. This avoids the issue.

The added test is for a different but related potential issue - if
signal is called while the task is running, we need to make sure we run
again immediately.

I've not added a test for the race, since I've refactored the code to
avoid it entirely (so cannot insert a test before the removersToComplete
is incremented but after wakeUp -- I've reversed the order of these).

Change-Id: I3371bf27e9c18269ec8791d0a790ac89b92071a2
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/226389
Reviewed-by: Faizan Alam <faizan.alam@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>