MB-66334: Fix LoadVBucket when Fusion uploader is disabled by default
Previously, Magma enabled the FusionUploader by default, which ensured
data was periodically synced to the logstore. With the uploader
now disabled by default, the test failed because the expected data was
no longer present on the FusionLogstore.

This fix manually starts the FusionUploader and triggers a sync to
restore the expected behavior.

Change-Id: I6d736db807e2615f5906055e059bdac72d4dcf84
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/226695
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Rohan Suri <rohan.suri@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>