MB-65281: Disable deduplication for all SystemEvent Items
Collection's has always required that we don't deduplicate
in the following case.

1: create(c1)
2: mutate(c1.k1)
3: drop(c1)

Deduplication could lead to the following

2: mutate(c1.k1)
3: drop(c1)

Create removed by deduplication leaving a mutation to a collection
which appears as if it was never created.

This is tested by:

CollectionsDcpParameterizedTest::test_dcp_create_delete

MB-65281 noted that scope events were handled differently which
raised some questions. The result is that the existing mechanism
for preventing deduplication can be improved - that is don't
create a new checkpoint before/after a system-event, just use the
CanDeduplicate mechanism (part of CDC).

1) A create/modify/drop (all which have the same key) cannot
deduplicate.

2) Checkpoint usage is much more effcient, for example the creation of
n collections can now all go in one checkpoint, rather than n
checkpoints.

Change-Id: I539ddd9891d7e1fded87bc69898e259634bcd673
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/226896
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>