MB-66553: Add trace span for cookie reschedule wait
Add a new trace span cover the time from when a background
thread tried to notify a cookie until the frontend thread
serviced that request. All notification happens as part of
draining the notification queue in the EventBase object,
which may see starvation if the queue is long and the
other event loop keep rescheduling itself.

Change-Id: I9506676718f67c51bf15a24f4425852b8245e755
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/226979
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Well-Formed: Restriction Checker