MB-66385: Reduce the number of notifications
Change the behavior of Connection::triggerCallback to only
schedule a callback if none will be generated automatically
due to data transfer. To make sure we don't have to wait until
the *entire* userspace buffer was sent before scheduling a
refill we utilize bufferevents watermark to trigger a
callback once we pass a multiplier of 64k transferred to
the kernel.

The patch moves the actual execution of a notified cookie
back to the full timeslice it has for executing command
and not as part of receiving the notification (as that
would potentially give the connection an unfare amount
of extra timeslices)

Change-Id: I041afccdb3c94d33b0b7c928b3b037d05c9647d7
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/227000
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Well-Formed: Restriction Checker