MB-66552: Allow up to 64 single write to the kernel
Bufferevent end up using writev to send data to the socket
and schedules 1 write attempt at the time to the socket.
Allow it to pass on 64k in a single write operation (as
that is closer to the max size of an IP packet anyway).
At the end of the day we want the data to be moved from
the userspace buffer over to the kernel buffer and sent
as fast as possible.

Change-Id: Ie37945debabe2cb79d7395495ba54a76866ac806
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/227113
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>