MB-64463: Limit the number of shards to 16 when max_vbuckets=128
Added a new limit on the number of shards, which is calculated by
max_vbuckets / max_num_shards_vbucket_factor. It is defined in this way
to support different max_vbuckets configurations.

New configuration parameter: max_num_shards_vbucket_factor=8 (default).

For max_vbuckets=1024, this sets the limit to 128 which coincides with
the existing hard-coded limit.

For max_vbuckets=128, this limits the shards to 16 now.

For the remaining configurations we've used in the past:
 - for max_vbuckets=64, limits shards to 8
 - for max_vbuckets=48, limits shards to 6

Change-Id: Ic2e86c961826cc2b00a8f8325f1624cac8dfb197
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/227344
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>