MB-66609: Clear VBucket::takeover_backed_up flag at takeover
The ActiveStream's takeover procedure might set the
VBucket::takeover_backed_up flag for putting the vbucket in a state
where it temporarily rejects frontend operations. That is for ensuring
that the takeover procedure completes eventually.

That flag is set based on how long the stream takes for processing the
various "in memory" takeover phases (ie TakeoverSend and TakeoverWait).
That's all diven by the dcp_takeover_max_time config param (default to
60 seconds).

If that flag is set then our logic misses to reset it when the
takeover completes. The consequence is that the vbucket is left in a
takeover-backup state once takeover completes. If the vbucket is
then reset to non-dead and traffic is enabled on it, then the
vbucket would keep rejecting operations forever.

The patch addresses also the case where a stream is prematurely ended
in the middle of a takeover procedure (eg Producer disconnection). If
the VBucket::takeover_backed_up flag was already set at stream-end, the
flag is cleared.

Change-Id: Ib79ccf17ec8b10ad5cfc735509d55501a503ee5a
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/227471
Reviewed-by: Jim Walker <jim@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>