MB-66612: Use lastReplicatedSeqno in seqno-advance condition
curChkSeqno is incorrect for use in the guarding and input into the
call to sendSnapshotAndSeqnoAdvanced within
ActiveStream::processItemsInner.

curChkSeqno is assigned to the seqno of the empty checkpoint meta
item when registration finds nothing in checkpoints for the vbucket.
This means curChkSeqno is assigned a "future" seqno.

Using curChkSeqno as the endSeqno parameter for calling
sendSnapshotAndSeqnoAdvanced can then poison the ActiveStream for
a later Monotonic exception. This is because the endSeqno value
will get assigned to lastReadSeqno, a Monotonic type.

When a real mutation is later processed it will take the "future"
seqno and then crash memcached when we assign lastReadSeqno again.

Change-Id: I714f1c043b99cf033b531355000272279801a62c
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/227492
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Well-Formed: Restriction Checker