MB-66628: SetClusterConfig CAS must be 0 [2/8]
ns_server never implemented the support for setting the control token
which means that it has always provided CAS == 0. Remove support for
the control token as the list of commands supporting it isn't very
consistent and a new task should be scheduled to do a full scrub over
all commands if we want to implement this.

Change-Id: I98a507f0ddd9d7125deb14064389d71d7e6ea207
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/227532
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Pavlos Georgiou <pavlos.georgiou@couchbase.com>