MB-66628: Pause and Resume must have CAS=0 [4/8]
ns_server never implemented the support for setting the control token
which means that it has always provided CAS == 0. Remove support for
the control token as the list of commands supporting it isn't very
consistent and a new task should be scheduled to do a full scrub over
all commands if we want to implement this.

Change-Id: Ic7dab10626d3f67214fee32a902136de8aa87bde
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/227533
Reviewed-by: Pavlos Georgiou <pavlos.georgiou@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>