MB-66628: SetParam must use CAS=0 [3/8]
ns_server never implemented the support for setting the control token
which means that it has always provided CAS == 0. Remove support for
the control token as the list of commands supporting it isn't very
consistent and a new task should be scheduled to do a full scrub over
all commands if we want to implement this.

Change-Id: I1d2e37620f8d26c3f3685d8a579374aefd6c4ae5
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/227526
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>