MB-66628: DelVbucket require CAS==0 [5/8]
ns_server never implemented the support for setting the control token
which means that it has always provided CAS == 0. Remove support for
the control token as the list of commands supporting it isn't very
consistent and a new task should be scheduled to do a full scrub over
all commands if we want to implement this.

Change-Id: Id17f78ae29f793f3d9002edc4cb108542c881522
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/227535
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@couchbase.com>