MB-66628: SetVbucket require CAS==0 [6/8]
ns_server never implemented the support for setting the control token
which means that it has always provided CAS == 0. Remove support for
the control token as the list of commands supporting it isn't very
consistent and a new task should be scheduled to do a full scrub over
all commands if we want to implement this.

Change-Id: I30ccd92965b564e0cafb03426a4a99e3aa199ede
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/227536
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>