MB-66628: Set/Get Ctrl Token is no longer supported [8/8]
ns_server never implemented the support for setting the control token
which means that it has always provided CAS == 0. Remove support for
the control token as the list of commands supporting it isn't very
consistent and a new task should be scheduled to do a full scrub over
all commands if we want to implement this.

Change-Id: Ibdbfb2a9a9e78561c4b791443da937cee326f59a
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/227538
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>