Refactor Magma read only unit test
A FileSystem is now considered initialised only if all of its member
fields are non-null. This unit test was only partly setting up the FS by
only setting the MakeDirectory member. This would now lead to Magma not
using the config provided FS and rather fallback to its internal FS.
Which is not what we want for this test since we're looking to simulate
a DiskFull.

Linked Magma change:
https://review.couchbase.org/c/magma/+/227676

Change-Id: I25914978e4d5b7fcd5cb546caa41797d5e5e06f4
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/227678
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>