MB-67040, MB-66731: Write the full header block [5/n]
Previously we would write the "header" and then seek (and possibly
create a hole in the file) for the rest of the header. We can might
as well just write full header and drop the system call. This will
make our life easier when we want to add support for an underlying
IO layer which would add encryption to the data.

This patch fix MB-67040 by using the same algorithm to locate
the first data block in the write pattern. The per-item max size
is set to 512 bytes, so don't allow alog_block_size to be less
than that

Change-Id: Ic32be46c55dfe58ad9dd9d62aec5eed4a35457f8
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/227900
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>