MB-66731: Open the mutation log read only to read [1/n]
Moving forward it shall not be possible to open an existing
file to append more data. The current usage in our system work
the way that it create a new "temporary" file; populate it
with all the information available before the file is closed
and renamed in place.

We had unit tests which didn't specify that the intended use
of the mutation log was "read only"

Change-Id: I9b066a7be244100e26c53aa453baa5d7294a7b6a
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/227896
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>