MB-65656: Create new fusion stat group 'migration'
This stat group exposes fusion migration progress stats:
TotalBytes and CompletedBytes. A stat call can either contain a
vbid to get the migration stats for a particular vbucket, or no
vbid in which case the migration stats for all vbuckets is returned.
An example of an aggregated Json response will be:
{
  "vb_0": {
    "completed_bytes": "100",
    "total_bytes": 1000,
  },
  "vb_1": ...
}

Change-Id: I25277cacca9a0bbde82c768cdeaf62e152000385
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/227980
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>