MB-66819: Remove support for V1-3 of access log
Remove support for old versions of the access log. These old
versions would be read as part of offline upgrade, and the
previous version of the access log was last used in 6.6 (which
was EOLd in October 2023).

One cannot upgrade directly to the current version of Couchbase,
which makes it desriable to kill the extra complexity added by
reading and writing different versions of the file.

An upgrade from 6.6 -> 8.0 would need to go through 7.2.3,
and one would need to be running 7.2.3 until a new access.log
file gets written if one would want the bucket to be warmed
up as part of the next upgrade.

Change-Id: I345b9370248bce0c9b6825ae3b4735d3fb2e825d
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/228033
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Faizan Alam <faizan.alam@couchbase.com>