MB-66731: Use FileReader in MutationLogReader [12/n]
Use cb::crypto::FileReader to read the access log. In this
version we'll read the entire file in one chunk as part of
creating the instance, and wrap the "pread" calls from
the iterator to return chunks from the block.

There is a refactor in platform which change the underlying
cb::crypto::FileReader to read the file chunk by chunk
and once thats merged the internals of this class will be
updated to utilize that functionality.

Change-Id: Ie1d9e1d7b7a26ccce3ae02e7ad43a02188ba8ee3
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/228273
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>