MB-66577: Send Snapshot Marker v2.2 only when hps in range
Consider the following example:

1. 3 node cluster: n1 (7.6), n2 (8.0), n3 (8.0)
2. vb:1, active: n1, replica: n3
3. n1 sends disk snapshot marker v2.0 (start seqno: 0, end seqno: 10),
   n3 moves hps to snapEnd (hps = 10).
4. rebalance out n1, n2 becomes active.
5. n2 send disk snapshot marker v2.2 (start seqno: 11, end seqno: 15,
   hps: 0)
6. n3 processes the above snapshot marker & tries to set hps = 0 &
   crashes since the monotonic invariant fails.

With this fix, we include the hps in the snapshot marker only if it's in
the [snapStart, snapEnd] range. That guarantees even if the replica
received a snapshot marker from a lower version node (& the hps is
bumped till last snapEnd seqno) before, when we process the new snapshot
marker the hps is monotonically increasing.

The test in dcp_reflection_test tests the same.

This is likely to happen only when the hps is initally set to 0
(no durability workload on the cluster) & we send out a snapshot marker
v2.2 from an 8.0 node - but I have made the check a little
more generic by making sure the HPS in within the snapshot range.

Change-Id: I114dadab7054302c59af711fc0128ae646a72819
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/228130
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>