MB-66920: Front end threads should not stop early
If a graceful shutdown happens while there is multiple connections
connected to the system the BucketDestroyer will try to run
a task in the front end context to try to disconnect all thread
specific connections. If the thread is stopped this will hang
"forever".

Historically (before using thread specific connection queues) we
wanted to terminate the front end threads as soon as possible (when
there was no connections bound to the thread) as it would free up
resources on the server; and when we arrived to the step to stop
the threads they would already be stopped. Given that the server
was in the process of stopping we could perform that check as part
of dispatching new connections. Now that we have the per-thread
connection queue the threads must run until all buckets are shut
down.

Change-Id: I912c12ab870436c764e945360a18c5ac82c45863
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/228307
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>