MB-66735: Change magma_fusion_namespace_prefix from a bucket config ...
to a static global configuration

Currently, accessing magma_fusion_namespace_prefix requires an instance
of a bucket configuration. However, there are two known scenarios where
we need to use this prefix without access to a bucket configuration:

    1) When validating the namespace passed into deleteFusionNamespace,
we need to ensure that the first part of the namespace begins with the
defined prefix.
    2) Magma requires KV to pass the prefix when issuing a
GetFusionNamespace command. However, because this is a non-bucket
command, we do not have access to the bucket configuration to retrieve
the prefix.

Both of these will be fixed in the comming patches.

Change-Id: I4e6196b6bd804deb878bf5e1c7dfef4d4c355bec
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/227835
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>