MB-66962: Let normal shutdown try to shut down all buckets...
... at the same time. Shutting down a bucket has to wait for
all clients to disconnect which may take some time so we should
initiate that on all buckets immediately.

Reapply the patch initially reviewed on
https://review.couchbase.org/c/kv_engine/+/228384

The patch was reverted as MB-66995 was filed the next day and
looked like the crash was during bucket shutdown. Reverting
the patch did not resolve that problem.

Change-Id: Ia7117226de72857e7f9a57e5f09f5b6de9745900
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/228544
Reviewed-by: Faizan Alam <faizan.alam@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>