Revert "MB-66962: Let normal shutdown try to shut down all buckets..."
From looking at the memcached log attached to MB-66995 it looks
like the node was running shutdown when it crashed (unfortunately
there is no minidump available, and the crashed thread is not
present in the stack trace on the console.

Given that it looks like the system crashed while doing shutdown
and the bug appeared close to when the change was merged it is
likely related. Revert while investigating

This reverts commit c01ffd9545e25825f6fc24e9e9d175a702ba2cc8.

Change-Id: Ia90dfd498d4cd26a2a876e7785aac1ac45c4ee2b
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/228499
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@couchbase.com>