MB-66731: Remove debug log message [17/n]
The "expected" use case would be to read the mutation log sequential
from beginning to end even if it supports random read from offsets.
While debugging an unrelated problem a log message was added for
the case where we would try to access an offset earlier than the
current position. This debug log message is no longer needed and
should be removed.

Change-Id: I3a6ef8d795a925697acd27a460ba00d334b908a7
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/229053
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>